Switch back to 16-bit memory architecutre, damn it's a huge mess

+++ include/lc3vmwin_cache.hpp
@@ -18,7 +18,7 @@ extern uint16_t cacheCount;
 
 extern struct lc3Cache codeCache[];
 
-struct lc3Cache cache_create_block(uint8_t memory[], uint16_t lc3Address);
+struct lc3Cache cache_create_block(uint16_t memory[], uint16_t lc3Address);
 void cache_clear();
 void cache_add(struct lc3Cache c);
 int cache_find(uint16_t address);
